embedding_search/requirements.txt
transformers>=4.51.3
einops==0.8.0
torch>=2.0.0
torchvision>=0.15.0
decord==0.6.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0
opencv-python>=4.8.0
pandas>=2.0.0
huggingface_hub>=0.23.0
pyyaml>=6.0.0
faiss-cpu>=1.7.4  # For CPU systems
# faiss-gpu>=1.7.4  # Uncomment for GPU systems with CUDA
pyarrow>=14.0.0  # For Parquet support
plotly>=5.17.0  # For interactive visualizations (as in official demo)
streamlit>=1.28.0  # Optional: for web interface like official demo
umap-learn>=0.5.4  # For UMAP dimensionality reduction (better than t-SNE)
trimap>=1.1.4  # For TriMAP dimensionality reduction (alternative to t-SNE)
# Additional visualization and ML dependencies
seaborn>=0.12.0  # Statistical data visualization

# Performance monitoring dependencies (optional)
psutil>=5.9.0  # System monitoring
GPUtil>=1.4.0  # GPU monitoring (optional)
nvidia-ml-py>=12.535.108  # NVIDIA Management Library (optional)
